Update brainstorming-datasets.md
I've moved a few of the FAQs to the main section describing the task as I think it's important that they read these rather than just going to the FAQ if they (think) they have a question.

I've left the "How will this be reviewed" section as is for now, but I think this will likely be revamped in a few weeks time. I, as a CD, do actually review the datasets because I want to make sure that a) they aren't using the iris dataset, b) they aren't using synthetic data because it's the "easier" option, and c) they understand the dataset well enough to make correct assumptions/statements/conclusions about it. For example, I was reviewing a course a few months ago that used BMI data and the values in the Female BMI column ranged from 100-130. BMIs typically range from 18-30, with 30 representing an obese human being. Values in the range listed are damn near record-breaking and were very unlikely to be correct. I want to make sure the instructor has the authority to determine whether his/her data is accurate/clean.